feat(ui): 第二轮 UI 美化，更紧凑精致的简约风格
- QSS: GroupBox 纯白底+淡灰边框+灰色小字标题，降低视觉重量
- QSS: CheckBox indicator 缩至 16px，行间距收紧
- QSS: Slider handle 缩至 14px，按钮统一 28px 高度
- QSS: 保存按钮去掉边框，更简洁
- QSS: 新增 QToolTip 样式（白底圆角）
- 布局: Tab 内 spacing 8→6，规则列表 spacing 2
- 布局: 窗口固定宽度 520px
- fix: CallNextHookEx lParam 溢出修复

+++ src/hotkey_manager.py
@@ -124,7 +124,7 @@ class HotkeyManager(QObject):
                     # 检查 Ctrl 是否按下（高位为1表示按下）
                     if user32.GetAsyncKeyState(VK_CONTROL) & 0x8000:
                         self._on_ctrl_c()
-            return user32.CallNextHookEx(None, nCode, wParam, lParam)
+            return user32.CallNextHookEx(None, nCode, wParam, ctypes.c_long(lParam))
 
         self._ll_proc = _HOOKPROC(hook_proc)
         self._ll_hook = user32.SetWindowsHookExW(
